cleaned up `mod.rs` files
