+++ tests/generic_full_path/mod.rs
@@ -1,10 +1,3 @@
-use utoipa::OpenApi;
-use utoipauto_macro::utoipauto;
-
 mod more_schemas;
 mod schemas;
-
-#[utoipauto(paths = "./tests/generic_full_path")]
-#[derive(OpenApi)]
-#[openapi(info(title = "Percentage API", version = "1.0.0"))]
-pub struct CrateApiDocs;
+mod test;
